Workaround for slow getTransform function
In Matlab R2020b getTransform is too slow (~ 7 ms per call)
to be usable in real time with many drones. For this reason
we use an additional Python node to publish position and
velocity of each crazyflie on topics called 'cf#/pos' and 'cf#/vel'.
The values in the latest messages on these topics are returned
by the position() and velocity() functions.